pyproject.toml
[tool.poetry]
name = "standings-cross"
version = "0.1.0"
description = "API for Standings Cross application (Test task)"
authors = ["Niyaz Validov <niyazvalidov@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.75.2"
uvicorn = "^0.17.6"
pydantic = "^1.9.0"
tortoise-orm = "^0.19.0"
asyncpg = "^0.25.0"
aerich = "^0.6.3"
python-multipart = "^0.0.5"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]
aiosqlite = "^0.17.0"
pytest = "^7.1.1"
pytest-asyncio = "^0.18.3"
black = "^22.3.0"


[tool.aerich]
tortoise_orm = "src.db.TORTOISE_ORM"
location = "./migrations"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
include = '\.pyi?$'
exclude = '/(\.eggs|\.git|\.hg|\.mypy_cache|\.nox|\.tox|\.venv|_build|buck-out|build|dist|\.pytest_cache)/'

[tool.isort]
profile = "black"
